refactor(api): move certificationCenterMembership serializer
to src/shared
